Revert "refactor: reorganiza features e aplica DRY em escalas"
This reverts commit b4aec39c322f3ccb2c8c420acddb9ce04a29370f.
