+++ frontend-app/src/components/pages/Dashboard/ModalServico.tsx
@@ -1,8 +1,8 @@
 import { Plus, Trash2, X } from "lucide-react"
-import { Button } from "../../../components/ui/Button"
-import { Input } from "../../../components/ui/Input"
-import { useModalServico } from "../hooks/useModalServico"
-import type { Aluno } from "../types/aluno.types"
+import { Button } from "../../ui/Button"
+import { Input } from "../../ui/Input"
+import { useModalServico } from "../../../hooks/useModalServico"
+import type { Aluno } from "../../../types/aluno.types"
 
 
 interface ModalProps {
