+++ frontend-app/src/hooks/useModalServico.tsx
@@ -1,6 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { buscarAlunos } from "../../../services/alunos.service";
-import { criarNovoServico } from "../services/servicos.service";
+import { buscarAlunos } from "../services/alunos.service";
+import { criarNovoServico } from "../services/sevicos.service";
 import { SARGENTO_DIA_ATUAL_QUERY_KEY } from "./useSargentoDiaAtual";
 import { useServicoFormState } from "./useServicoFormState";
