+++ frontend-app/src/services/sevicos.service.ts
@@ -1,4 +1,4 @@
-import { api } from "../../../services/api";
+import { api } from "./api";
 
 interface PayloadNovoServico {
   data: string;
