+++ frontend-app/src/types/alterecao.types.ts
