+++ frontend-app/src/types/aluno.types.ts
